add debug log lines for soap messages send/receive
