+++ algosec/api_clients/fire_flow.py
@@ -14,7 +14,8 @@ Examples:
 """
 import logging
 
-from algosec.api_clients.base import SoapAPIClient, report_soap_failure
+from algosec.api_clients.base import SoapAPIClient
+from algosec.helpers import report_soap_failure
 from algosec.errors import AlgoSecLoginError, AlgoSecAPIError
 
 logger = logging.getLogger(__name__)
